+++ ChangeLog.rst
@@ -18,6 +18,8 @@ Unreleased Changes
 * Incorporated several patches from the FreeBSD port. libfuse should
   now compile under FreeBSD without the need for patches.
 
+* The passthrough_ll example now supports writeback caching.
+
 libfuse 3.1.0 (2017-07-08)
 ==========================